+++ lib/extensions/uint8list_extension.dart
@@ -4,14 +4,13 @@ import 'package:asn1lib/asn1lib.dart';
 import 'package:crypto/crypto.dart';
 import 'package:flutter/foundation.dart';
 
+import 'general_type_extension.dart';
+
 extension Uint8ListExtension on Uint8List {
   String? asCertificateDerGetSpkiPin() {
     try {
       final foundSpki = _getSpki(this);
-      if (foundSpki != null) {
-        return _getSpkiPin(foundSpki);
-      }
-      return null;
+      return foundSpki?.let((it) => _getSpkiPin(it));
     } catch (ex) {
       debugPrint(ex.toString());
       return null;